refactor: add package handlermapping
