+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handlermapping/AnnotationHandlerAdapter.java
@@ -1,4 +1,6 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.handlermapping;
+
+import nextstep.mvc.tobe.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
