+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handlermapping/HandlerMapping.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.tobe.handlermapping;
 
 import javax.servlet.http.HttpServletRequest;
